Implement tracing fallback to log files
The following implements a transition period where tracing is still
sent to the log files via the same setup. This can be turned on via
setting the tracing to anything but level.ClientNone.

The implementing just wraps the driver.Driver implementation, by
pushing the logging tracer on the context for the tracing context to
peel off the span.

Again, this isn't required, but does help with transitioning away
from the older style method.